+++ CHANGELOG.md
@@ -102,6 +102,7 @@
 - Stock Rules : Add massUpdate on 'Stock location', 'Product', 'Refill type' and 'Min. Qty'.
 - Inventory : Add filter on two new string fields on form when click on 'Fill inventory' button.
 - Add direct order stock location.
+- Inventory:Allow to do an inventory for one product.
 
 ## Improvements
 - Move project resource menu entry to project configuration menu.